Refactor: Remove unused ManuellFordelingsgrunnlag data classes and update imports
