+++ gradle/libs.versions.toml
@@ -4,7 +4,7 @@ komponenter = "2.0.129"
 tilgang = "1.0.234"
 behandlingsflyt = "0.0.644"
 oppgave = "0.0.207"
-arenaoppslag = "0.5.35"
+arenaoppslag = "0.6.0"
 resilience4j = "2.4.0"
 kafka = "4.3.1"
 logback = "1.6.1"
